delta-warehouse/requirement.txt
delta-spark        3.3.0
importlib_metadata 8.7.0
pip                25.1.1
py4j               0.10.9.7
pyspark            3.5.6
zipp               3.23.0